refactor: move `isValidSystemType` utility to foundry module
