+++ module/utils.mjs
@@ -1,5 +1,3 @@
-import { SYSTEM_TYPES } from "./constants/constants.mjs";
-
 // region Foundry
 
 /**
@@ -275,29 +273,6 @@ export function replaceFormulaData( formula, data, { actor, item, missing="0", p
 export function staticStatusId( status ) {
   if ( status.length >= 16 ) return status.substring( 0, 16 );
   return status.padEnd( 16, "0" );
-}
-
-// endregion
-
-// region System
-
-/**
- * Check whether the provided system type is valid for the given document type,
- * or in general if no document type is provided.
- * @param {string} systemType The system type to check.
- * @param {string} [documentType] The document type to check against.
- * @returns {boolean} True if the system type is valid, false otherwise.
- */
-export function isValidSystemType( systemType, documentType ) {
-  if ( documentType && !Object.keys( SYSTEM_TYPES ).includes( documentType ) ) return false;
-
-  const validTypes = documentType
-    ? Object.values( SYSTEM_TYPES[ documentType ] )
-    : Object.values( SYSTEM_TYPES ).map(
-      types => Object.values( types )
-    ).flat();
-
-  return validTypes.includes( systemType );
 }
 
 // endregion
